refactor (OmokRule): 룰 규칙 상태 개선
