+++ src/main/kotlin/omok/domain/gameState/BlackWin.kt
@@ -11,6 +11,9 @@ class BlackWin(
     override val isRunning: Boolean = false
 
     override fun play(point: OmokPoint): GameState {
-        throw IllegalStateException()
+        throw IllegalStateException(ERROR_ALREADY_FINISHED)
+    }
+    companion object {
+        private const val ERROR_ALREADY_FINISHED = "이미 검은 돌이 이겼기 때문에 더 이상 진행할 수 없습니다."
     }
 }
