+++ src/main/kotlin/omok/domain/gameState/WhiteWin.kt
@@ -11,6 +11,10 @@ class WhiteWin(
     override val isRunning: Boolean = false
 
     override fun play(point: OmokPoint): GameState {
-        throw IllegalStateException()
+        throw IllegalStateException(ERROR_ALREADY_FINISHED)
+    }
+
+    companion object {
+        private const val ERROR_ALREADY_FINISHED = "이미 흰 돌이 이겼기 때문에 더 이상 진행할 수 없습니다."
     }
 }
