--- src/main/kotlin/omok/domain/omokRule/OmokRuleInterface.kt
@@ -1,5 +1,0 @@
-package omok.domain.omokRule
-
-interface OmokRuleInterface {
-    fun validate(board: List<List<Int>>, position: Pair<Int, Int>): Boolean
-}
