+++ src/main/kotlin/omok/domain/omokRule/WhiteWinRule.kt
@@ -1,12 +1,12 @@
 package omok.domain.omokRule
 
-object WhiteWinRule : OmokRule(WHITE_STONE, BLACK_STONE) {
+object WhiteWinRule : OmokRuleImpl(WHITE_STONE, BLACK_STONE) {
     override fun validate(board: List<List<Int>>, position: Pair<Int, Int>): Boolean =
         directions.map { direction -> checkBlackWin(board, position, direction) }.contains(true)
 
     private fun checkBlackWin(board: List<List<Int>>, position: Pair<Int, Int>, direction: Pair<Int, Int>): Boolean {
-        val direction2 = direction.let { (dx, dy) -> -dx to -dy }
-        val (stone1, blink1) = search(board, position, direction2)
+        val oppositeDirection = direction.let { (dx, dy) -> Pair(-dx, -dy) }
+        val (stone1, blink1) = search(board, position, oppositeDirection)
         val (stone2, blink2) = search(board, position, direction)
 
         return when {
